# api/fomo-swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Endpoints labeled open do not require authentication. The protected endpoints on the other hand, do require\nauthentication. I'm not sure how long the jwt token should last. I'm thinking we should set the expire on\nthat token to be super short - like 5 minutes (upto an hour maybe?) to minimize the amount of time an\nattacker can use that token. The refresh token will last longer - currently 7 days. If you make a request\nto a protected endpoint using a \"Refresh\" token in your request headers, you will receive a new\nauthorization token (set-authorization) and refresh token (set-refresh) in the response headers when\nyou make a request with an expired authorization token. You MUST replace both tokens in your request headers\nto stay authenticated. The old refresh token gets replaced on the backend therefore, you need to use the\nnew refresh token to remain actively logged in.",
    "title": "FOMO API",
    "version": "0.0.1"
  },
  "host": "stage.fomo.exchange",
  "basePath": "/api",
  "paths": {
...
    "/keys": {
      "get": {
        "description": "The api secrets will not be returned in the response data.",
        "tags": [
          "keys"
        ],
        "summary": "get all user keys (protected)",
        "operationId": "GetAllKey",
        "responses": {
          "200": {
            "description": "\"data\" will contain a list of key info with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseKeyListSuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      },
      "post": {
        "description": "Associate a new exchange api key to a user's account. Secrets will not be returned in response data.",
        "tags": [
          "keys"
        ],
        "summary": "add an api key (protected)",
        "operationId": "AddKey",
        "parameters": [
          {
            "x-go-name": "Exchange",
            "description": "Required.",
            "name": "exchange",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Key",
            "description": "Required.",
            "name": "key",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Secret",
            "description": "Required.",
            "name": "secret",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Description",
            "description": "Optional.",
            "name": "description",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\"data\" will contain key info with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseKeySuccess"
            }
          },
          "400": {
            "description": "missing params with \"status\": \"fail\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    },
    "/keys/:keyID": {
      "get": {
        "description": "Gets a user's key by the key ID. The secret will not be returned in the response data.",
        "tags": [
          "keys"
        ],
        "summary": "get a key (protected)",
        "operationId": "GetKey",
        "responses": {
          "200": {
            "description": "\"data\" will contain key stuffs with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseKeySuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      },
      "put": {
        "description": "The user can only update the description of an added key. The secret will not be returned.",
        "tags": [
          "keys"
        ],
        "summary": "update a user api key (protected)",
        "operationId": "UpdateKey",
        "parameters": [
          {
            "x-go-name": "Description",
            "description": "Required.",
            "name": "description",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\"data\" will contain key info with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseKeySuccess"
            }
          },
          "400": {
            "description": "missing params with \"status\": \"fail\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      },
      "delete": {
        "description": "This will remove the api key from the system.",
        "tags": [
          "keys"
        ],
        "summary": "remove user api key (protected)",
        "operationId": "DeleteKey",
        "responses": {
          "200": {
            "description": "data will be null with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseKeySuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Key": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange": {
          "type": "string",
          "x-go-name": "Exchange"
        },
        "key": {
          "type": "string",
          "x-go-name": "Key"
        },
        "keyID": {
          "type": "string",
          "x-go-name": "KeyID"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "KeyList": {
      "type": "object",
      "properties": {
        "keys": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Key"
          },
          "x-go-name": "Keys"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "ResponseKeyListSuccess": {
      "type": "object",
      "title": "A ResponseKeyListSuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/KeyList"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
    "ResponseKeySuccess": {
      "type": "object",
      "title": "A ResponseKeySuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/UserKeyData"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "UserKeyData": {
      "type": "object",
      "properties": {
        "key": {
          "$ref": "#/definitions/Key"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "responseError": {
      "description": "This response may or may not include data encapsulating the user information.",
      "type": "object",
      "title": "A ResponseSuccess will always contain a status of \"successful\".",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "ResponseError",
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}